Move empty_parenthesized into the parentheses.rs
